Calculate distance_traveled

+++ micro_ros_raspberrypi_pico_sdk/hardware/Motor.cpp
@@ -52,11 +52,11 @@ void Motor::setSpeed(int speedPWM) {
     currentPwm = speedPWM;
 }
 
-float Motor::getSpeed() {
+double Motor::getSpeed() {
     // Use the IRSensor to calculate the rotations per second
     int number_counts_per_rev = irSensor->getSpeed();
     int radius = 0.05;
-    float distance_traveled = 2 * 3.14159 * radius * number_counts_per_rev;
+    double distance_traveled = 2 * 3.14159 * radius * number_counts_per_rev;
 
     return distance_traveled;
 }
